// generated/src/app/api-models/benefit-package-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class BenefitPackageDetail {
  benefitPackageId: string;
  seqBenPackage: number;
  benefitRule: string;
  startDate: Date;
  endDate: Date;
  parProvReq: string;
  authReq: string;
  riderReqFilter: string;
  supDepRule: string;
  noMosContFrom: number;
  noMosContThru: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  processingOrder: number;
  riderOperator: string;
  adjustForOcPaid: string;
  parOperator: string;
  authOperator: string;
  applyPar: string;
  authLevel: string;
  applyRider: string;
  weightedValueAccum: string;
}